# docs/openapi.yaml
openapi: 3.1.0
info:
  title: API Carte UAC
  version: 1.0.0
  description: |
    API de la carte du campus de l'Université d'Abomey-Calavi.
    Toutes les réponses JSON suivent la même enveloppe : `{ data, meta? }` en cas de succès,
    `{ error: { code, message, status } }` en cas d'échec. Les listes sont toujours paginées (`meta`).
    Ce fichier est le contrat de référence : toute modification de l'API commence ici.
  license:
    name: MPL-2.0
    url: https://mozilla.org/MPL/2.0/
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    adminSession:
      type: apiKey
      in: cookie
      name: uac_admin_session
      description: Cookie httpOnly posé par POST /admin/session (appli web).
    adminToken:
      type: http
      scheme: bearer
      description: Jeton ADMIN_TOKEN, réservé aux scripts et à l'intégration continue.
    contributorCookie:
      type: apiKey
      in: cookie
      name: uac_contributor
      description: Cookie httpOnly posé par POST /contributors, valable 180 jours. Identifie un téléphone contributeur.
  parameters:
    RequestId:
      name: X-Request-ID
      in: header
      required: false
      description: Identifiant de traçage (8 à 64 caractères alphanumériques ou tirets). Généré s'il est absent, renvoyé dans la réponse.
      schema: { type: string }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Order:
      name: order
      in: query
      schema: { type: string, enum: [asc, desc], default: asc }
...
    ProposalId:
      name: proposalId
      in: path
      required: true
      schema: { type: string }
    ContributorId:
      name: contributorId
      in: path
      required: true
      description: Identifiant d'un contributeur (« me » est réservé au téléphone qui appelle).
      schema: { type: string }
...
  schemas:
    Position:
      type: object
      required: [longitude, latitude]
      properties:
        longitude: { type: number, minimum: -180, maximum: 180 }
        latitude: { type: number, minimum: -90, maximum: 90 }
    Entrance:
      allOf:
        - $ref: '#/components/schemas/Position'
        - type: object
          properties:
            note: { type: string, maxLength: 120 }
    PlaceInput:
      type: object
      required: [name, longitude, latitude]
      properties:
        name: { type: string, maxLength: 120 }
        category:
          type: string
          enum:
            [
              administration,
              lecture-hall,
              classroom,
              faculty,
              library,
              food,
              health,
              sport,
              housing,
              worship,
              service,
              transport,
              gate,
              landmark,
              other,
            ]
          default: other
        aliases:
          type: array
          maxItems: 15
          items: { type: string, maxLength: 60 }
        description: { type: string, maxLength: 1000 }
        access: { type: string, maxLength: 300, description: 'Étage, salles, horaires.' }
        longitude: { type: number }
        latitude: { type: number }
        entrances:
          type: array
          maxItems: 10
          items: { $ref: '#/components/schemas/Entrance' }
...
    PathInput:
      type: object
      required: [coordinates]
      properties:
        type: { type: string, enum: [footpath, road, track, corridor, stairs], default: footpath }
        name: { type: string, maxLength: 120 }
        isFloodProne: { type: boolean, default: false }
        coordinates:
          type: array
          minItems: 2
          maxItems: 5000
          items:
            type: array
            prefixItems: [{ type: number }, { type: number }]
            description: '[longitude, latitude]'
...
    PaginationMeta:
      type: object
      required: [page, limit, total]
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, status]
          properties:
            code: { type: string, example: PLACE_NOT_FOUND }
            message: { type: string, example: Lieu inexistant }
            status: { type: integer, example: 404 }
...
    DevicePosition:
      type: object
      required: [longitude, latitude, accuracy, ageMs]
      description: Position du téléphone au moment de l'envoi. Seule sa précision est conservée.
      properties:
        longitude: { type: number, minimum: -180, maximum: 180 }
        latitude: { type: number, minimum: -90, maximum: 90 }
        accuracy: { type: number, minimum: 0, maximum: 50, description: 'Précision annoncée, en mètres.' }
        ageMs: { type: number, maximum: 120000, description: Âge de la position selon l'horloge du téléphone. }
    ProposalInput:
      type: object
      required: [entityType, action, payload, devicePosition]
      properties:
        entityType: { type: string, enum: [place, path] }
        action: { type: string, enum: [create, update, report], description: update est réservé aux lieux. }
        targetId: { type: string, maxLength: 64, description: Obligatoire pour update et report. }
        payload:
          oneOf:
            - $ref: '#/components/schemas/PlaceInput'
            - $ref: '#/components/schemas/PathInput'
            - type: object
              required: [message]
              properties:
                message: { type: string, maxLength: 500 }
        devicePosition: { $ref: '#/components/schemas/DevicePosition' }
    OwnContributor:
      type: object
      properties:
        id: { type: string }
        pseudonym: { type: [string, 'null'], maxLength: 40 }
        status: { type: string, enum: [new, trusted, blocked] }
        createdAt: { type: string, format: date-time }
    OwnProposal:
      type: object
      properties:
        id: { type: string }
        entityType: { type: string, enum: [place, path] }
        action: { type: string, enum: [create, update, report] }
        targetId: { type: [string, 'null'] }
        payload: { type: object }
        status: { type: string, enum: [pending, accepted, rejected, withdrawn] }
        reviewNote: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        reviewedAt: { type: [string, 'null'], format: date-time }
    Contributor:
      type: object
      description: Contributeur vu par un relecteur, avec son bilan.
      properties:
        id: { type: string }
        linkId: { type: string }
        pseudonym: { type: [string, 'null'] }
        status: { type: string, enum: [new, trusted, blocked] }
        createdAt: { type: string, format: date-time }
        lastSeenAt: { type: string, format: date-time }
        acceptedCount: { type: integer }
        rejectedCount: { type: integer }
        pendingCount: { type: integer }
    Proposal:
      type: object
      description: Proposition vue par un relecteur.
      properties:
        id: { type: string }
        contributorId: { type: [string, 'null'], description: Vide si le contributeur a été supprimé. }
        contributor:
          type: [object, 'null']
          properties:
            id: { type: string }
            pseudonym: { type: [string, 'null'] }
            status: { type: string, enum: [new, trusted, blocked] }
        entityType: { type: string, enum: [place, path] }
        action: { type: string, enum: [create, update, report] }
        targetId: { type: [string, 'null'] }
        targetUpdatedAt: { type: [string, 'null'], format: date-time }
        payload: { type: object }
        positionAccuracyMeters: { type: number }
        status: { type: string, enum: [pending, accepted, rejected, withdrawn] }
        reviewerKind: { type: [string, 'null'], description: admin ou reviewer ; vide si publiée directement. }
        reviewerId: { type: [string, 'null'] }
        reviewNote: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        reviewedAt: { type: [string, 'null'], format: date-time }
    ProposalDetail:
      allOf:
        - $ref: '#/components/schemas/Proposal'
        - type: object
          properties:
            contributor:
              oneOf:
                - $ref: '#/components/schemas/Contributor'
                - type: 'null'
            target: { type: [object, 'null'], description: Version actuelle du lieu ou du chemin visé. }
            hasConflict:
              {
                type: boolean,
                description: "Vrai si la cible a changé ou disparu depuis l'envoi, pour une proposition en attente.",
              }
    MapChange:
      type: object
      properties:
        id: { type: integer }
        entityType: { type: string, enum: [place, path, map] }
        entityId: { type: [string, 'null'] }
        action: { type: string, enum: [create, update, delete, revert, bulk] }
        beforeState: { type: [object, 'null'] }
        afterState: { type: [object, 'null'] }
        actorKind: { type: string, enum: [admin, reviewer, contributor, command] }
        actorId: { type: [string, 'null'] }
        actorName: { type: [string, 'null'], description: Nom du relecteur ou pseudo du contributeur. }
        proposalId: { type: [string, 'null'] }
        revertsChangeId: { type: [integer, 'null'] }
        createdAt: { type: string, format: date-time }
...
  responses:
    BadRequest:
      description: Données invalides (codes VALIDATION_ERROR, INVALID_JSON, INVALID_PAGE, INVALID_LIMIT, INVALID_SORT, INVALID_ORDER, INVALID_CATEGORY, INVALID_PATH_TYPE, INVALID_FORMAT, CONFIRMATION_REQUIRED, INVALID_STATUS, INVALID_ENTITY_TYPE, INVALID_CONTRIBUTOR_ID, INVALID_ENTITY_ID).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Session absente ou expirée (UNAUTHORIZED), ou jeton incorrect (INVALID_TOKEN).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Ressource inexistante (PLACE_NOT_FOUND, PATH_NOT_FOUND, ROUTE_NOT_FOUND, NO_PUBLIC_LINK, LINK_NOT_FOUND, TARGET_NOT_FOUND, PROPOSAL_NOT_FOUND, CONTRIBUTOR_NOT_FOUND, CHANGE_NOT_FOUND, REVIEWER_NOT_FOUND).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Proposition déjà traitée (PROPOSAL_ALREADY_REVIEWED), élément modifié depuis (CHANGE_OUTDATED) ou opération en masse (CHANGE_NOT_REVERTIBLE).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    TooManyRequests:
      description: Trop d'envois dans l'heure (TOO_MANY_REQUESTS).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotAContributor:
      description: Ce téléphone n'a pas rejoint la contribution (NOT_A_CONTRIBUTOR).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Téléphone bloqué (CONTRIBUTOR_BLOCKED), ou action réservée à l'administrateur (FORBIDDEN).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    PresenceNotVerified:
      description: Présence sur le campus non vérifiée (POSITION_REQUIRED, OUTSIDE_CAMPUS, POSITION_INACCURATE, POSITION_TOO_OLD, GEOMETRY_OUTSIDE_CAMPUS).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unavailable:
      description: Contributions suspendues (CONTRIBUTIONS_PAUSED) ou périmètre du campus non configuré (PERIMETER_NOT_CONFIGURED).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
security: []
paths:
...
  /contributors:
    get:
      tags: [Relecture]
      summary: Liste des contributeurs, avec leur bilan
      security: [{ adminSession: [] }, { adminToken: [] }]
      parameters:
        - { $ref: '#/components/parameters/RequestId' }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
        - name: status
          in: query
          schema: { type: string, enum: [new, trusted, blocked] }
      responses:
        '200':
          description: Page de contributeurs, du plus récent au plus ancien
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Contributor' } }
                  meta: { $ref: '#/components/schemas/PaginationMeta' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Contribution]
      summary: Rejoindre la contribution par un lien
      description: Dépose le cookie uac_contributor. Un téléphone déjà inscrit reçoit son contributeur existant (200), quel que soit le lien.
      parameters: [{ $ref: '#/components/parameters/RequestId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [linkCode]
              properties:
                linkCode: { type: string }
                pseudonym: { type: string, maxLength: 40 }
      responses:
        '200':
          description: Téléphone déjà inscrit
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: '#/components/schemas/OwnContributor' } } }
        '201':
          description: Contributeur créé (en-tête Set-Cookie)
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: '#/components/schemas/OwnContributor' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '410':
          description: Lien fermé (LINK_CLOSED)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
  /proposals:
    get:
      tags: [Relecture]
      summary: File des propositions
      security: [{ adminSession: [] }, { adminToken: [] }]
      parameters:
        - { $ref: '#/components/parameters/RequestId' }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Order' }
        - name: status
          in: query
          schema: { type: string, enum: [pending, accepted, rejected, withdrawn] }
        - name: entity-type
          in: query
          schema: { type: string, enum: [place, path] }
        - name: contributor-id
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Page de propositions, de la plus ancienne à la plus récente par défaut
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Proposal' } }
                  meta: { $ref: '#/components/schemas/PaginationMeta' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Contribution]
      summary: Envoyer une proposition
      description: |
        Contributeur « nouveau » : la proposition attend une relecture (pending). Contributeur « de confiance » :
        elle est publiée aussitôt (accepted), sauf un signalement, toujours relu. La position du téléphone est
        vérifiée à chaque envoi ; seule sa précision est conservée.
      security: [{ contributorCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/RequestId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProposalInput' }
      responses:
        '201':
          description: Proposition enregistrée
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: '#/components/schemas/OwnProposal' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/NotAContributor' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/PresenceNotVerified' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /proposals/{proposalId}:
    parameters: [{ $ref: '#/components/parameters/ProposalId' }, { $ref: '#/components/parameters/RequestId' }]
    get:
      tags: [Relecture]
      summary: Détail d'une proposition, version actuelle de la cible et conflit éventuel
      security: [{ adminSession: [] }, { adminToken: [] }]
      responses:
        '200':
          description: Proposition
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: '#/components/schemas/ProposalDetail' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Relecture]
      summary: Accepter (publier) ou refuser une proposition
      description: Accepter un signalement le marque comme traité, sans modifier la carte.
      security: [{ adminSession: [] }, { adminToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [accepted, rejected] }
                note: { type: string, maxLength: 300, description: Montrée au contributeur. }
      responses:
        '200':
          description: Proposition traitée
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: '#/components/schemas/Proposal' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /contributors/{contributorId}:
    parameters: [{ $ref: '#/components/parameters/ContributorId' }, { $ref: '#/components/parameters/RequestId' }]
    patch:
      tags: [Relecture]
      summary: Accorder ou retirer la confiance, bloquer un téléphone
      description: Bloquer refuse aussitôt tout ce que le téléphone a en attente.
      security: [{ adminSession: [] }, { adminToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [new, trusted, blocked] }
      responses:
        '200':
          description: Contributeur modifié
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: '#/components/schemas/Contributor' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /map-changes:
    parameters: [{ $ref: '#/components/parameters/RequestId' }]
    get:
      tags: [Relecture]
      summary: Historique des modifications, de la plus récente à la plus ancienne
      security: [{ adminSession: [] }, { adminToken: [] }]
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
        - name: entity-type
          in: query
          schema: { type: string, enum: [place, path, map] }
        - name: entity-id
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Page de l'historique
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/MapChange' } }
                  meta: { $ref: '#/components/schemas/PaginationMeta' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Relecture]
      summary: Annuler une modification
      description: L'élément retrouve son état d'avant. Refusé s'il a changé depuis (annuler d'abord les modifications plus récentes) ou pour une opération en masse.
      security: [{ adminSession: [] }, { adminToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [revertsChangeId]
              properties:
                revertsChangeId: { type: integer, minimum: 1 }
      responses:
        '201':
          description: Annulation inscrite à l'historique
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: '#/components/schemas/MapChange' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...